permanent status box
