+++ src/app.tsx
@@ -2,101 +2,8 @@ import * as React from "react"
 const blessed = require("blessed")
 const { render } = require("react-blessed/dist/fiber/fiber")
 const authorize = require("./authorize")
+import { GmailMessage, GmailThread } from "./gmail-classes"
 import { inspect } from "util"
-const { openURL } = require("./utils")
-
-const USE_TRELLO_DESKTOP = true
-
-class GmailThread {
-  _thread: any
-  messages: [GmailMessage]
-  snippet: string
-  id: string
-
-  isOpen = false
-
-  constructor(thread) {
-    this._thread = thread
-
-    this.messages = thread.messages.reverse().map(message => new GmailMessage(message))
-    this.snippet = this._thread.snippet
-    this.id = thread.id
-  }
-}
-
-class GmailMessage {
-  _message: any
-  _headers: object
-  payload: object
-
-  constructor(message) {
-    this._message = message
-    this._headers = {}
-    this._headers = message.payload.headers.reduce((memo, header) => {
-      memo[header.name.toLowerCase()] = header.value
-      return memo
-    }, {})
-    this.payload = message.payload
-  }
-
-  get id() {
-    return this._message.id
-  }
-
-  get threadId() {
-    return this._message.threadId
-  }
-
-  get subject() {
-    return this._headers["subject"]
-  }
-
-  get plainText() {
-    const { parts } = this._message.payload
-
-    if (!parts) return ""
-
-    const plainText = parts.find(p => p.mimeType === "text/plain") || parts[0]
-
-    if (!plainText.body.data) return ""
-
-    return new Buffer(plainText.body.data, "base64").toString("utf8")
-  }
-
-  get externalURL() {
-    return this.githubURL || this.trelloURL
-  }
-
-  get githubURL() {
-    const match = this.plainText.match(/github:\s+(http.+)/im)
-
-    return match && match[1]
-  }
-
-  get trelloURL() {
-    const match = this.plainText.match(/\((https:\/\/trello.com.+?)\)/)
-
-    if (!match) return null
-
-    const originalURL = match[1]
-
-    if (USE_TRELLO_DESKTOP) {
-      return originalURL.replace("https://", "trello://")
-    } else {
-      return originalURL
-    }
-  }
-
-  open({ background = false } = {}) {
-    return new Promise((resolve, reject) => {
-      if (this.externalURL) {
-        resolve(openURL(this.externalURL, { background }))
-      } else {
-        reject("no external URL")
-      }
-    })
-  }
-}
 
 interface IAppProps {
   gmail: {
@@ -114,6 +21,7 @@ interface IAppState {
   error: any
   lastArchivedThreadId: number | null
   openThreads: { [threadId: string]: boolean }
+  status: string | null
 }
 
 class App extends React.Component<IAppProps, IAppState> {
@@ -126,7 +34,8 @@ class App extends React.Component<IAppProps, IAppState> {
       openThreads: {},
       selectedIndex: null,
       error: null,
-      lastArchivedThreadId: null
+      lastArchivedThreadId: null,
+      status: null
     }
   }
 
@@ -141,7 +50,7 @@ class App extends React.Component<IAppProps, IAppState> {
   }
 
   reloadInbox = () => {
-    this.setState({ error: "loading" })
+    this.logStatus("loading")
     const { gmail } = this.props
     const userId = "me"
     return gmail.threads
@@ -152,12 +61,15 @@ class App extends React.Component<IAppProps, IAppState> {
       .then(({ threads }) => {
         return Promise.all(
           (threads || []).map(thread => {
-            return gmail.threads.get({ userId, id: thread.id })
+            return gmail.threads.get({ userId, id: thread.id }).then(thread => {
+              this.logStatus(`loaded ${thread.id}`)
+              return thread
+            })
           })
         )
       })
       .then(threads => threads.map(t => new GmailThread(t)))
-      .then(threads => this.setState({ threads, error: null }))
+      .then(threads => this.setState({ threads }))
   }
 
   handleMessageListMovement = key => {
@@ -196,7 +108,10 @@ class App extends React.Component<IAppProps, IAppState> {
     const { messages } = this
     const selectedMessage = messages[messageList.selected]
     if (full === "C-o") {
-      selectedMessage.open({ background: true }).catch(this.logError)
+      selectedMessage
+        .open({ background: true })
+        .then(() => this.logStatus(`open ${selectedMessage.externalURL}`))
+        .catch(this.logError)
     } else if (full === "C-d") {
       this.archiveThread(selectedMessage.threadId)
     } else if (full === "C-z" && this.state.lastArchivedThreadId) {
@@ -210,18 +125,21 @@ class App extends React.Component<IAppProps, IAppState> {
         return { openThreads }
       })
     } else if (full === "h" || full === "left") {
-      this.setState(({ openThreads }) => {
-        delete openThreads[selectedMessage.threadId]
+      this.setState(
+        ({ openThreads }) => {
+          delete openThreads[selectedMessage.threadId]
 
-        return { openThreads }
-      }, () => {
-        const thread = this.state.threads.find(t => t.id == selectedMessage.threadId)
-        if (!thread) throw "no thread!"
+          return { openThreads }
+        },
+        () => {
+          const thread = this.state.threads.find(t => t.id == selectedMessage.threadId)
+          if (!thread) throw "no thread!"
 
-        const message = thread.messages[0]
-        messageList.select(this.messages.findIndex(m => m.id == message.id))
-        messageList.screen.render()
-      })
+          const message = thread.messages[0]
+          messageList.select(this.messages.findIndex(m => m.id == message.id))
+          messageList.screen.render()
+        }
+      )
     }
   }
 
@@ -253,6 +171,14 @@ class App extends React.Component<IAppProps, IAppState> {
 
   logError = error => {
     this.setState({ error })
+  }
+
+  private statusTimeout: NodeJS.Timer | null = null
+
+  logStatus = status => {
+    this.setState({ status })
+    this.statusTimeout && clearTimeout(this.statusTimeout)
+    this.statusTimeout = setTimeout(() => this.setState({ status: null }), 2000)
   }
 
   get messages(): GmailMessage[] {
@@ -289,7 +215,7 @@ class App extends React.Component<IAppProps, IAppState> {
   }
 
   render() {
-    const { error } = this.state
+    const { error, status } = this.state
     const { messages, messageSubjects } = this
     const selectedMessage = messages[this.state.selectedIndex || 0]
 
@@ -306,7 +232,10 @@ class App extends React.Component<IAppProps, IAppState> {
             this.setState({ selectedIndex: index })
           }}
           onSelect={(_item, index) => {
-            messages[index].open().catch(this.logError)
+            messages[index]
+              .open()
+              .then(() => this.logStatus(`open ${selectedMessage.externalURL}`))
+              .catch(this.logError)
           }}
           onKeypress={this.handleMessageListKeypress}
           ref={ref => (this.messageList = ref)}
@@ -315,8 +244,10 @@ class App extends React.Component<IAppProps, IAppState> {
           border={{ type: "line" }}
           style={{ border: { fg: "blue" }, selected: { bg: "gray" } }}
           top="20%"
-          height={error ? "60%" : "80%"}
+          height={(error ? "60%" : "80%") + "-2"}
           width="100%"
+          mouse
+          scrollable
         >
           {selectedMessage && selectedMessage.plainText}
         </box>
@@ -324,7 +255,7 @@ class App extends React.Component<IAppProps, IAppState> {
           <box
             border={{ type: "line" }}
             style={{ border: { fg: "red" } }}
-            top="80%"
+            top="70%"
             height="20%"
             width="100%"
             mouse
@@ -333,6 +264,16 @@ class App extends React.Component<IAppProps, IAppState> {
             {inspect(error)}
           </box>
         )}
+        <box
+          border={{ type: "line" }}
+          style={{ border: { fg: "gray" } }}
+          bottom="0"
+          height="0%+3"
+          width="100%"
+          scrollable
+          tags
+          content={status || ""}
+        />
       </element>
     )
   }
